// src/modules/users-security/shared/types/auth.types.ts
export interface LoginRequest {
  email: string;
  password: string;
}

export interface FuncionInfo {
  id_funcion: number;
  nombre: string;
  modulo: string;
  nivel_acceso: string;
}

export interface RolInfo {
  id_rol: number;
  nombre: string;
}

export interface AuthUser {
  id_usuario: number;
  email: string;
  estado: string;
}

export interface LoginResponse {
  accessToken: string;
  user: AuthUser;
  rol: RolInfo;
  funciones: FuncionInfo[];
}

export interface AuthState {
  isAuthenticated: boolean;
  isLoading: boolean;
  user: AuthUser | null;
  rol: RolInfo | null;
  funciones: FuncionInfo[];
}